lambda/requirements.txt
pynacl
pymysql